UI: Move plugin table related logic to plugin table
